+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Adds the Champ de Bataille server.
+- New game notifications for players.
 
 ## [v20.6.2](https://github.com/lexicalunit/spellbot/releases/tag/v20.6.2) - 2026-05-27